#143. renamed component according to new business logic
