+++ src/pages/tools.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import ToolkitFeed from "../components/shared/ToolkitFeed"
+import ToolAppsFeed from "../components/shared/ToolAppsFeed"
 import BaseLayout from "../components/static/BaseLayout"
 import MainHeading from "../components/static/MainHeading"
 
@@ -11,7 +11,7 @@ const Toolkit = () => {
       >ReInHerit Toolkit</MainHeading>
       <p className="w-75">All tools on the platform are <b style={{color:"#d7487e"}}>categorized</b> and public available. </p>
       <br></br>
-      <ToolkitFeed></ToolkitFeed>
+      <ToolAppsFeed></ToolAppsFeed>
     </BaseLayout>
   )
 }
